Creación de representante con base de datos

+++ src/main/java/com/natame/api/negocio/entidades/Representante.java
@@ -37,5 +37,5 @@ public class Representante {
     private String diraccion;
     @NotNull
     @Positive
-    private byte codigoRegion;
+    private int codigoRegion;
 }
